feat: new cta image scroller and marquee to improve performance
